Stats library: Generic Template for stats collection
This patch defines a generic template that can be used for metrics/stats
collection. Typical data of interest includes CPU util, memory util,
error stats, etc.

lib/sys_stats.rb is a common file and it defines two methods
to collect and parse stats at a given interval. The methods
signatures are shown below:
 start_collecting_system_stats(metrics, interval=-1)
 stop_collecting_system_stats(parsers)

The 'metrics' and 'parsers' arguments in the methods above are Hash
objects, each hash element's value is a function pointer. The logic to
collect and parse the data is implemented in the functions addressed by
the function pointers in 'metrics' and 'parsers' respectively.

It is recommended that a Module file, that defines functions to collect
and parse common metrics, be defined for each vertical (LSP, BSP, etc.)
A sample implementation of such a Module file is provided with this
patch in LSP/metrics.rb

A sample usage of this generic template is also provided in
LSP/default_target_test.rb

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>