System suspend resume performance test.
This patch implements the verification of system health after resuming from suspend.
In this application, the system will be configured to suspend and resume after
sleeping for few seconds. But the system  checks for reliability before suspend and after resuming.
The way this reliability is being verified  in this application is as follows:
1)  before suspending the system,  application plays either media or graphics
or both depending on the test case and verifies performance parameters.
2) if things goes well, the system is put to suspend and configure to wake
up after few minutes using alarm.
3) if the system wakes up then step one is repeated.
This test goes for a number of iteration configured in the test case.

Signed-off-by: Yebio Mesfin <ymesfin@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>