LSP: new script to run ltp scenario files
This is very similar to the existing dev_test2.rb script currently
use to run ltp tests, the only difference is that this one provides
more information about the number of tests run and identifies the
failed tests in the test result notes.
This is useful in cases where a single Testlink test is running many
ltp tests.

The patch also increases the timeout when printing the test sdtout
logs because the logs can now be significantly bigger than before

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>