Automate UBoot Nand ECC tests.
This test introduce N bits error in each 512 chunk in one page;
and it verify if those N bits errors can be corrected in UBoot.

The result message looks like below.
For cnt=0: Original byte is 0xa7, err_byte is 0x58 and the corrected_byte is 0xa7;
For cnt=1: Original byte is 0x31, err_byte is 0xce and the corrected_byte is 0x31;
For cnt=2: Original byte is 0x74, err_byte is 0x8b and the corrected_byte is 0x74;
For cnt=3: Original byte is 0x30, err_byte is 0xcf and the corrected_byte is 0x30;

Signed-off-by: Yan Liu <yan-liu@ti.com>