Test application to perform usb device tests in parallel with usbhost tests
This patch modifies an existing test application for usb gadget to run
usb host tests in parallel. A test case parameter will determine whether
such a parallel test is to be run as well as indicate what type of host
test to be run. There are also some fixes to make the original application
more robust by checking the usb interface on host PC dynamically
rather than assume it to be usb0.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>