LSP/default_test_module: Add option to install user_bins
New software asset usr_bins can be used to pass a tarball with
binaries or just a single binary file that should be installed
in the root filesystem prior to running the tests.

The typical use case for this option is when a tester have a
dependency on an application that it is not part of the filesystem.
The new logic will install (i.e. untar) the file into ~/bin/
directory and it will add such directory to the PATH.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>