am335x-evm boot order automation script.
The file defines functions used to test boot order uart->spi->nand
on am335x EVM. Since the target is boot order related to nand,
only uart->fail over->nand is tested here. Booting from spi is not
being tested here. Once the platform is booted from uart and image is
loaded to nand, the platform is expected to fail over to nand, if spi
is not programmed.

Signed-off-by: Yebio Mesfin <ymesfin@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>