Linux: default_test_module: Avoid tftp race condition
It is possible to have two or more TEEs of the same type running in
the same PC and using the same software asset. In such scenarios,
one TEE could start running a test while a file is still being
copied to the tftp server directory by another TEE.
This patch makes file path unique by relaying on the TEE ID which
should be unique. The new path is short enough and it is no longer
necessary to create a softlink.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>