Linux: Add new script to set performance governor
In the past, the test cases that measured drivers performance did
not set the performance governor but instead used whatever was the
default kernel governor. The current default governor in the kernel
is the ondemand governor which could affect drivers performance.
This patch makes sure that peformance governor is selected prior to
running the tests.

The drivers performance test cases in Testlink will be modified to
use this script.

Signed-off-by: Carlos Hernandez <ceh@ti.com>