Remove lib/result_forms.rb
This patch removes the windowing library lib/results_forms.rb
from the repo. This library was moved to vatf sources
because vatf cleans the ruby files required by test scripts
but it does not clear the Classes and constants created by
them, this causes an error when loading the windowing library
again on a new test case.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>