scmcsdk:enhanced ethernet performance application
This update is to fix the issue with this application not measuring the
full bandwidth of the 1Gig interface. Also fixed is the static way that
the ethernet interfaces were previously being handled.

Fixes include using the threading option of Iperf, fixing an intermittent
issue with the binary search routine and adding the ability to specify
a minimum throughput mbps as a pass/fail criteria. Made the handling of
the ethernet interfaces more dynamic so that a broad range of ethernet
ports can be specified. Static IP addresses for the non-dhcp enabled
interfaces are derived from the static_ip_base_address base specified in
the bench file. If static_ip_base_address is set to blank ('') then
dhclient will be used to obtain an IP address.

Three files were modified; ipsec_connect_module.rb, ipsec_tests.rb and
utilities.rb. The files are located in scmcsdk/syslib area of the VATF scripts
repository.

Signed-off-by: Elvita Lobo <elvita@ti.com>