Set NS result for tests with wrong kernel config
There are a set of kernel config options that were classified as
'optional' in ltp-ddt. Previous behavior was to mark as failure tests
where kernel config option was not set as expected by the test.
However, there are valid use cases where a required config option is
not set (e.g. disable debug option to improve performance in
production builds), and in such cases it is better to mark test as
skipped instead of marking as failure.

Signed-off-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>