scmcsdk:updated ethernet performance application
This update is to fix the issue where sometimes the application
can not read the througput measurements from Iperf and to add
support for jumbo frames.

Removed waitfor a specific command and went back to the waitfor
that will cause the send command to timeout. Letting the send
command timeout is much more effective in capturing the performance
data rather than waiting for a specific command. Added the command
to cat the /proc/net/snmp log. Set last ethernet dhcp port to 0 to
match the latest MCSDK release. Added set mtu size method along with
a find ethernet interface by IP address method to get the PCs ethernet
interface to which to add the jumbo frame mtu size.

Two files were modified; ipsec_tests.rb and utilities.rb. The files are
located in scmcsdk/syslib area of the VATF scripts repository.

Signed-off-by: Elvita Lobo <elvita@ti.com>