scmcsdk:added large file tftp nand test to armtest
This update is to add file TFTP support to armtest framework.

Added new TFTP methods to armtest framework to support
file TFTP specified in an armtest test case. Corrected an issue
in the armtest where it did not always detect error messages if
they were not in the first line of the response (Line 118). Added
informational messages for when a test case failed.

Two files were modified; common_utils.rb and run_arm_test.rb. The files
are located in scmcsdk/syslib area of the VATF scripts repository.

Signed-off-by: Elvita Lobo <elvita@ti.com>