Allow the user to specify the number of boot attempts
This patch implements a feature that allows the user to specify
the maximum number of times the default_test_modules should try
to succesfully boot the board. Before this patch the module tried
to boot the board once and raises an exception if the board did
not boot. After this patch the user can specify the number of times
the module will try to boot the board succesfully via var_boot_attempts
variable that can be set in them command line with -a or in the build
description. The number of boot attempts is set to 1 by default.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>