LSP: Add timestamping test script
This script is intended to be generic enough to test any timestamping
application, but is primarily written to work with the kernel timestamping
test application. See

https://www.kernel.org/doc/Documentation/networking/timestamping/timestamping.c

for more details

Inputs to the script are the logfile with timestamps, resolution of offset
error the user wants to report, timestamping interval and regular
expression to find timestamps in the logfile. The script will report
offset errors (difference between observed and expected timestamps)
if they are larger than resolution value. This will catch clock roll over,
if logfile has sufficient timestamp values. The script will also report
if clock drift is observed.

Signed-off-by: Elvita Lobo <elvita@ti.com>