scmcsdk:added connection ping check & fixed mtu size
This update is to fix an issue where the test keeps running even
though the connection between the PC and EVM can not pass data.
Since the automatic throughput detection can not rely on every
rate tested to finish this pre-check avoids wasting time trying
to figure out the throughput when the connection will not pass
data. As a part of this fix the mtu size was also fixed and
is included because it is a small fix in the same code section.

Signed-off-by: Elvita Lobo <elvita@ti.com>