scmcsdk:added ability to load platform specific ipsecmgr
This update adds the ability to load the correct ipsecmgr apps by
using the platform name specified in the bench file.

Starting with MCSDK 3.1.0 the ipsecmgr daemon and cmd shell are
platform specific and require the platform name to be part of
the command. This patch adds code needed to fetch and use the
platform name from the bench file for ipsecmgr apps.

Signed-off-by: Elvita Lobo <elvita@ti.com>