scmcsdk:increased throughput rate resolution
This update is to increase throughput rate resolution.

Previous automatic throughput determination resolution was 25Mbps
which meant the rate obtained would be a multiple of 25Mbps. This
worked fine for high throughput rates but not so well for low
throughput rates. Changed the resolution to 2Mbps which works
well for both high and low throughput rates.

Signed-off-by: Elvita Lobo <elvita@ti.com>