Power: Specify multilmeter connection type in bench
This patch allows the user to specify the type of connection used
to communicate with the multimeter in the bench file. Before this
patch the connection type used in the script was hardcoded to serial
after this patch the user can use either serial or tcp_ip by setting
the conn_type parameter in the params bench entry of the multimeter.
If no conn_type parameter is specified in the bench the connection
type defaults to serial.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>