scmcsdk:fixed intermittent inflow setup issue
This update fixes an intermittent failure to establish the
IPSEC inflow connection. This was caused by using the default
send command response. There are times when the default command
response does not allow the complete list of enviroment variables
to be displayed and because of that triggers using certain variables
were not triggered. Send command response now waits for the env
command as listed at the end of the list so that the entire list
will be displayed before the response is evaluated.

Signed-off-by: Elvita Lobo <elvita@ti.com>