scmcsdk:updated binary search routine logic
Currently the binary search routine will return zero during a search
session, but this is not a valid throughput test specification. The
result of the test maybe zero but the bandwidth tested should be
greater than zero. The binary search routine now returns the search
step specified if the search result is less than the search step.

Signed-off-by: Elvita Lobo <elvita@ti.com>