scmcsdk:removed throughput measuring limitation for 10G
This update removes a hard coded limit for 10G of 1.4G to keep Iperf
from hanging during testing. Since the Iperf process has for awhile
now been removed and restarted for each measurement this hard coded
value is no longer required and needs to be removed to facilitate
the much higher throughput measurements needed for jumbo frames.
Also included in this patch and related to testing 10G is increasing
the maximum allowed number of attempts to determine the bandwidth. In
this same section a variable name misspelling was corrected as well.

Signed-off-by: Elvita Lobo <elvita@ti.com>