scmcsdk:enhanced ipsec setup application for rekey
This enhancement allows the user through the test case to specify the
ipsec.conf parameters that control rekey. The default rekey parameters
are set as they were before to rekey after 48 hours. Also in this patch
is the new stop_offload command usage since it is trivial and is
located in the same area of code as the rekey additions.

Signed-off-by: Elvita Lobo <elvita@ti.com>