LSP/power_resume_time.rb: handle partial responses
The script did not handle well the case where after trying to wakeup
the board from UART, the board did not completely resume and the
complete resume log is split in two separate send commands.
This change appends responses for each commands to resume the board
into the same string, which is used to parse the resume time info.

Signed-off-by: Carlos Hernandez <ceh@ti.com>