Update to fix intermittently failing throughput results
This update allows the auto bandwidth detection method to continue
determining the max bandwidth if the throughput test fails to produce
a valid result. On rare occasions, the determined bandwidth will work
only for the 5 second detection time but will not work for the full
test time. This update allows the auto detection to continue when this
condition occurs, but this time it will use the full test time to
determine the max bandwidth starting just below the amount of the
failed bandwidth. This fix should help to eliminate test runs that fail
to produce a throughput result while still saving time by using the
quick auto detection method the majority of the time. In fixing this
issue it made sense to move the execution of the auto detect method to
the PerfUtilities class where it is already located. The auto detection
and performance test are now done in a single call instead of two
separate calls.

Signed-off-by: Elvita Lobo <elvita@ti.com>