Update to fix process_running? method from always returning true
This update fixes the process_running? method so that it will properly
return the running state of a process. Before the fix it would always
return true due to the fact that the command line which is part of the
response string contained a digit and this would satisfy the match
criteria. The other issue was that if there was a digit in the EVM's
prompt such as "K2HK_EVM:" then it would also match since the prompt
is also part of the response.

Signed-off-by: Carlos Hernandez <ceh@ti.com>