Update to fix intermittent failure of performance test cases
This update fixes a send receive sequence issue that was caused by not
properly waiting for the command in progress to finish before starting
the next command. The fix is to wait for the command prompt as well as
the expected string of the output from the executed command to make
sure the command is complete before continuing to the next command.

Signed-off-by: Elvita Lobo <elvita@ti.com>