Fix for process_running? returning false on Linux PC side
This update fixes the process_running? method so that it will properly
return the running state of a process on both the DUT and Linux PC
sides. Due to differences in the way that the response is done
between the DUT and the Linux PC, using the prompt for validation
does not work for the Linux PC side. This fix checks for formatting
that is unique to the "ps aux" command to validate the existence of
a process. Fix has been verified on both DUT and Linux PC side.

Signed-off-by: Yan Liu <yan-liu@ti.com>