Added generic method for verifying a command ran successfully
Added generic method check_cmd? that runs a command and verifies
that the command ran successful. The success of a command is based
on the return value of that command. A return value of 0 indicates
that the command ran successfully and a non-zero return value means
that the command was not successful. The check_cmd? method also
handles the difference in the way that the command return value is
retrieved for a local host vs a remote host. This method has been
verified on both DUT and Linux PC side for negative and positive
scenarios.

Signed-off-by: Yan Liu <yan-liu@ti.com>