scmcsdk: loading ipsec manager module before IPSEC start
This update allows the IPSEC connection utility to load the
ipsecmgr_mod.ko module before starting any IPSEC connections. This
was done to comply with the user documentation for MCSDK as well as
the expert opinion of the IPSEC developer that a user should not rely
on this module being automatically loaded, it should always be
intentionally loaded before establishing IPSEC connections.

Signed-off-by: Elvita Lobo <elvita@ti.com>