scmcsdk: added the ability to have IPSEC use secure store
This update adds the ability to establish an IPSEC connection using
the credentials kept in a TI secure store. To facilitate initializing
the secure store it was necessary to modify the smart_cmd_send method
to have the ability to disable the by default check of the echo back
of the command in order to properly check for responses. To fix the
current bug in the strongswan.conf file, a method was added to correct
the file so that secure store can be tested by automation before the
small correction is made by the developers.

Signed-off-by: Elvita Lobo <elvita@ti.com>