scmcsdk:fixed corner case issue loading ipsecmgr module
This update fixes an issue where the method checking and loading the
ipsecmgr_module could see multiple instances of the module and try to
load all of them. This issue would normally only show up when running
a friendly and multiple copies of a module exist on the EVM's file
system. The fix is to call the existing method that locates and
isolates the first occurrence of the file found so the calling method
can then load only that one occurrence. In the process of doing this
fix I have enhanced the get_file_location method to use the ability
of find to specify the location or locations of where you want to scan
for the file. This change allows the caller to specify more than one
scan path in order of precedence and avoids the need to change
directories to have find search the right path. The default scan
location is the top level directory.

Signed-off-by: Elvita Lobo <elvita@ti.com>