libs: Add signal handler to result forms
This patch adds a signal handler to the ResultWindow class included
in the result_forms library. Before this patch the window would
remain opened after the parent process was killed, requiring a user
to manually close it. After this patch the result window closes
when the parent process is killed. This will help avoid locking of
resources that may happen if semi-automated test that use this
lib are accidentally trigger and then the cancel button is pressed

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>