kernel_boot_test: Detects errors during reboot
It is possible that some drivers will trigger errors during the
shutdown process during board reboot, so try to catch such errors
and fail the test if they are seen.
Poking into /var/log/messages or /dev/kmsg is tricky during shutdown
and the error messages that we are trying to catch are not written
there because filesystem is already unmounted by the time they
happen. So our best choice is to rely on console messages.
Unfortunately prinkt console messages does not include the severity
level and a patch to included it was NAK [1]. So this patch looks for
timestamp signature followed by undesired keywords, such as fail and
error.

[1]: http://dflvsu0004.dal.design.ti.com/patch/11673/

Signed-off-by: Carlos Hernandez <ceh@ti.com>