A-Audio: Fix sync issue for multicard setups
This patch fixes a sync problems in boards/systems with multiple
audio cards. Before this patch the play/rec operations synced on
the last play/rec, which caused the test to returned early if the
last audio operation failed. This patch adds a wait pid call so
that the test wait until all audio operation end before returning

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>