uboot: Clean up uboot script/table used to test uboot commands
User need add new entries in the uboot table in order to run uboot tests on new platforms even
the values are the same as old platform. This patch modified the table to set default value if
there is any so new platform can also use it.
Also cleanup the table and remove un-used entries.

Signed-off-by: Yan Liu <yan-liu@ti.com>