boot_order_test: Change the way to set bootloader for usbspl
This patch 1) change the way to set bootloaders for usbspl;
Before, the script replace the name in dhcpd.conf with the to-be-tested
image names. This old way has problem when different platforms use the
same uboot binaries. Now this patch will not modify the names in dhcp config;
instead, it modifies the image name to match the expected name in dhcpd.conf.

2) Move platform specific functions to platform-specific script

Signed-off-by: Yan Liu <yan-liu@ti.com>