Added wrapper to run the GLSDK framework
The GLSDK framework is a separate test framework used by the
automotive team. The GLSDK framework is retrieved by cloning a git
repository to a directory accessible to the EVM. The GLSDK framework
files are modified for the platform to be used and then the framework
script is executed on the EVM to run the modules(test suites) specified.

This wrapper allows through the test case the setting of the module
and or modules to run, the max time the GLSDK framework should
allow for the modules to run, disabling or enabling of looping, and
disabling or enabling of run once feature. Please see header at the
top of the glsdk_utils.rb file for more details.

This implementation currently only supports NFS due to the fact that
the AM57xx does not currently support git or git package installation.
Using git directly on the EVM will be supported at a later time once
I have experience with installing and using the EVM's git package.

Signed-off-by: Calvin Wagner <cwagner@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>