LSP/A-GLSDK fixed timeout for gen of large report
A send_cmd timeout was happening during the GLSDK report generation
when there was a large report being created because of many test
failures.

The fix was to increase the time allowed for the report generator to
complete. Changed the timeout from 2 minutes to 15 minutes. Did some
testing of the report generation and for a heavily failed test run
with 92% failures, it took 122 seconds to complete the report.

Signed-off-by: Calvin Wagner <cwagner@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>