add support to disable / enable weston and matrix gui at system boot
Several testcases (like HDMI, Graphics) rely on the display port.

If weston and the Matrix GUI are started at boot, on AM5, the display port
is in use and these tests cannot be run.

This patch disables/enables the inittime scripts that enable weston and Matrix GUI

if var_disable_weston is set to 1, weston and matrix are disabled at boot

if var_disable_weston is set to 0, or not set at all, weston and matrix are
re-enabled (if previously disabled in the fs) at boot or not enabled at all
(if the fs does not enable them by default)

Signed-off-by: Elvita Lobo <elvita@ti.com>