uboot_nandecc.rb: Removed the dependency of 'saveenv' to nand for nandecc tests
This script was using 'saveenv' to get the nand test address; but, sometimes,
there is no 'saveenv' to nand support and this script could not be used for
uboot nandecc tests. Now, this patch removed this dependency and use one page
in the file-system partition. Before the test, this patch writes some random
data to this nand page and then uses this page to run nandecc tests.

Signed-off-by: Yan Liu <yan-liu@ti.com>