mcsdk: Enhanced perf data spreadsheet creation files
The current performance data parsing script does not have the support
for 10G ports tested. The instructions on how to capture the perf data
from TestLink and create the spreadsheets is kind of high level.

This fix adds parsing support for the 10G ports tested as well as adds
more detail to the instructions on how to capture the data and create
the spreadsheets. Also did some cleanup on the parsing script and some
display enhancements to the perf_gen.bat file.

Signed-off-by: Calvin Wagner <cwagner@ti.com>
Signed-off-by: Elvita Lobo <elvita@ti.com>