Acquire/Release Staf Semaphore Mutex Locks
Adds function which acquires staf semaphore lock, yields to
test code and releases the lock. Can be used for all test applications
which need exclusive use of a resource on the host machine or a
common test equipment.
Example uses are for: iperf tests, usb device tests and audio/video
tests which need exclusive access to test equipment.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>