ccs: multicore_loader support for ROV and all platforms
This patch reworks multicore_loader to support running the coredump
utility to read ROV data.  The loader now accepts any generic binary
and uses a lookup table in multicore_data to determine which cores to
load.  Some of the changes needed in the loader to facilitate the
coredump require changes in the bench file, so a sample bench has been
provided.

Signed-off-by: John Godbey <j-godbey@ti.com>
Signed-off-by: Elvita Lobo <elvita@ti.com>