Utilize semaphore locks to protect iperf tests and fix a bug in multiple ethernet interface tests
Uses staf_mutex function to ensure that test runs only on one tee
instance at a time on a given host machine. This will make sure
ethernet performance measurements are reliable. Test application
waits for an "iperf" lock to be acquired for 4 minutes before timing
out. Considering most of the tests running in parallel are of 1
minute duration, this ensures that the lock can be acquired even if
three tests are started at the same time within a reasonable wait.
This wait time might need to be fine-tuned based on results from
automated daily test runs.
There was a bug in which iperf was getting killed inside the
loop for multiple interface tests instead of outside the loop.
Fixing it here.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>