A-Video: Add auto option to vpe and wb test
This patch adds the option to run vpe and wb tests
in auto mode. When this option is enabled the test
is run and the processed video file is compared
against a golden reference to determine pass/fail.
To enable this option the test case must include
an "auto" parameter in the DUT parameters field

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>