A-GLSDK: Set is_clone_git_repo_on_pc dynamically
This patch adds code to determine the value of variable
is_clone_git_repo_on_pc dynamically. This variable
is used to decide where should a clone of the
glsdk test framework should occur. Before this patch
the value of is_clone_git_repo_on_pc was hardcoded to true
and the user had to change it manually if the dut was not
booting with an nfs file system. After this patch the value
of is_clone_git_repo_on_pc is determined based on wether
the bootargs used an nfs root device or not

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>