Add PPS (packets per second) data to perf_data for ethernet tests
Computes and adds pps data to perf_data for easier viewing and
tracking of performance.
Also, adds some conditions to ensure parsing and perf_data updates
are being done only for udp test cases which are successful.
The existing parsing logic is not valid for tcp tests and failed tests.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>

V2: Removing commented out line of code for clarity.
Signed-off-by: Yan Liu <yan-liu@ti.com>