A-Video: VPE load result_forms only for non-auto type test
This patch makes loading of the result_forms library
conditional depending on the type of test. Before
this patch result_forms was loaded regardless if the
test was automated or not, after this patch the
lib will only get loaded if test is not of type
"Automated". This changes allows wb automated test
to run on setups not configured for interactive tests

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>