KLP: Catch exceptions to help debugging failures
The method to check for Known Linux Problems sometimes
generated an exception if there were any problems that
would create nil dut objects. This make debugging the
failure harder since the KLP exception masks the original
exception that caused the dut object to be nil. This
patch fixes the problem by adding code to catch all
exceptions in the KLP method so that the original
failure cause or exception can be observed

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>