A-Video: Automate DSS-WB M2M negative tests
This patch adds code to automated DSS-WB M2M negatives.
The patch modifies the wb test script to check for a
new test case parameter negative_test. When this parameter
is specified in the test case the script will run the
WB M2M operation specified and then is will run an ls
command to verify that the system has not hang

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>