A-Display: Make result_forms loading conditional
This patch make loading of result_forms conditional on
test parameter auto. If auto is defined in the test case
the forms will not be loaded. Before this patch
result_forms was being required regardless of the type
of test, which resulted in an exception when the automated
version of the test was run in system without the wx
libraries

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>