wb: Determine number of frames using file size
The number of frames processed by v4l2-ctl that are
included in the file also contain some of the frames located
in the buffers when the count specified was reached which
make the total number of frames in the file the count
specified plus additional frames in the buffer. After
several runs it has been observed that the number of
frames included from the buffers is not constant.
This patch changes the way the number of frames used
to determine pass/fail from a fix number based on
the number of test frames specified to a runtime
calculation based on the size of the test file

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>