utils: Add function to utilize free system memory
This patch adds a new function use_memory that allows
the user to utilize part of the free memory available
before starting an app. It takes a parameter that
defines the number of bytes from free memory that will
not be utilize by the function. The memory utilization
is accomplished by making a call to memtester

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>
Signed-off-by: Carlos Hernandez <ceh@ti.com>