A-Display: Fix long timeout problem in drm tests
During negative test cases setting an unsupported mode was
taking too long to time out, since the expected pattern was
always looking for the fps to be printed out. This patch
changes this behavior by mixing the pattern passed with
known patterns that are always printed out by the sequence
of commands used. After this patch the negative test case
execution time is equal to the positive test cases

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>