AM5 IPC Testcase Automation
This patch impacts the IPC AM5 GP EVM nightly testcases. Given the implementation of PRSDK-1124, the IPC test images are available in ti-rtos-test-image and the rtos-bin file creation using top level makefile is deprecated. So the IPC automation needs to use the ti-rtos-test-image and the updated directory structure to pick up the binary images. The new directory structure does not have a platform token. Also, ROV files are available at the expected location in the new directory structure, so they do not need to be copied there in automation. The script was missing loggerbug for DSP and IPU cores, they were added. The IPU pass criteria was incorrect, it has been corrected in this patch. Also, AM5 needs a system reset. This is a new feature that was added for AM5. Testcases were failing due to IPC version changes, so this patch makes the code IPC version agnostic.
Please see below for a list of changes needed to support all the aforementioned updates.

multicore_data.rb

1. Updated new directory structure.
1. Removed platform_token construct due to new directory structure.
3. Added logger buf for DSP cores.
4. Added smp logger buf for IPU1 and IPU2 cores.
5. Added IPU output pass criteria.

multicore_loader.rb

1. Updated base_dir so that script is agnostic to ipc version.
2. Post processing ipc version string extracted from tarball.
3. Removed rtos_bins from directory structure due to new directory structure.
4. Remove prepending of base directory due to directory structure change.
5. Added system reset option, yes or no
6. Removed copy of rov files to binary directory due to new directory structure.

Signed-off-by: Vrinda Khetan <vrinda@ti.com>
Signed-off-by: Swapnil Mali <x0259461@ti.com>