A-Video: Refactor test to handle unused CRTCs being off
Now that CRTCs are turned off if not in use, the app has
to handle turning it so that the primary plane is allocated.
This was caused the test to fail when modetest was used
since both the primary plane and overlay would end up using
the same plane. This patch refactors the test to be based on
kmstest, which enables the crtcs (if disabled) before using
it

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>