A-Video: Automate capture test
This patch automates the sensor capture test. Before this patch
the test required user input to determine pass/fail. After this
patch pass/fail is determined automatically based on golden files.
The patch requires adding a 'sensor_box' => "<box id>" entry to
the dut params field in the bench file so that appropiate golden
files are used during the test

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>