A-Video: Set skipped frames based on reported fps
This patch changes the way the number of skip frames are
calculated in capture operations from resolution based
to reported fps. Since during capture skipping frames is
done to delay 10 seconds so the sensor can converge in
the capture settings. Changing to fps based calculations
will be make the delay more accurate

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>