A-Display: Improve result determination
This patch refactors the way in which driver messages are
classified to determine pass/fail. Before this patch all
error messages and kernel warnings/panics would have the
same priority. After this patch a kernel warning/panic
have highest priority and fails the test; "Invalid argument"
and "Not enough bandwidth" are classified as negative tests
and passed; and any other error messages are failures

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>