A-Video: Update vpe test based on new app version
This patch updates the VPE test script to work with the
latest version of the test app. In the latest version of
the test app the name of the application changed from
testvpe to test-v4l2-m2m and the dev node is no longer
hard coded to /dev/video0, but passed as a parameter
instead. This patch takes care of these changes

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>