default_test_module.rb: Set default 'fs_dev' to 'eth' if fs is defined
Since the patch ff286bc corrected 'fs_src_dev' to 'fs_dev' in
system_loader, the default 'fs_dev' should not be always set to
'mmc' anymore. In some cases like ramfs, the 'fs_dev' should be
set to 'eth'. So this patch set default 'fs_dev' to 'eth' if 'fs'
is passed.

Signed-off-by: Yan Liu <yan-liu@ti.com>