Adding the IoT Gateway automated test case
This test case clones the tidep0084 git repository, adds the
AWS certificates, adds the awsConfig.json file, adds a hosts
file to use the ENS requested firewall hole, and then pushes
all the files to the BBB in the test farm and runs the IoT
Gateway demo. If a successful connection to the AWS server is
made then this test is considered as passing.

Signed-off-by: Jason Reeder <jreeder@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>