Android: Refactor android base scripts
This patch updates android base scripts in order to allow
preparing a board for android testing based on .img image
files. Prior to this patch android validation efforts
were based on flashing a board with the new images and
then running the test on the board. This patch, together
with the correspoding vatf patches, allow flashing the
board (if needed) before running the tests. The patch
also includes updates to boot_time.rb that was used to
validate the changes in the patch

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>