LSP: Convert to string before checking asset
This patch converts to string the value returned
by the params Hash for an asset key, in the comparison
performed inside copy_sw_assets_to_tftproot. Before
this patch the check was performed against the value
returned by the Hash which would generate an exception
when trying to copy an asset if the value returned is
nil. The fix in this patch converts the value returned
to string which would skip copying being that the string
value of nil is ''

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>