Fix egress throughput and cpu utilization calculation logic
Egress throughput was not being measured correctly when using a dut port which
is in subnet other than that of the server PC's primary ethernet port which
is advertised as telnet_server in bench file. This patch gets the correct
server IP by using routing table output of PC so that throughput tests run on
non-eth0 will report correct values.
While at it:
1) Fix error in cpu utilization parsing logic
2) Integrate crypto threshold configuration settings to optimize IPSec
throughput (enable software fallback).

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>